requirements.txt
requests
pandas
streamlit
matplotlib
praw
python-dotenv
